+++ experiments/async_benchmark.py
@@ -10,6 +10,7 @@ from multiprocessing import Process, Value
 import random as r
 from time import sleep,time
 from torille import ToribashControl
+import argparse
 import sys
 
 NUM_JOINTS = 22
@@ -20,6 +21,16 @@ WARM_UP_SECONDS = 10
 # How many seconds will benchmark last
 BENCHMARK_SECONDS = 60
 
+parser = argparse.ArgumentParser()
+parser.add_argument("--warmup_time", default=WARM_UP_SECONDS,
+                    type=int, help="Seconds before benchmark starts (default: 10s)")
+parser.add_argument("--benchmark_time", default=BENCHMARK_SECONDS,
+                    type=int, help="How many seconds benchmark lasts (default: 60s)")
+parser.add_argument("num_instances", type=int, help="How many instances will be launched")
+parser.add_argument("match_frames", type=int, help="How many frames matches last")
+parser.add_argument("turn_frames", type=int, help="Amount of frames between turns")
+parser.add_argument("engagement_distance", type=int, help="How far apart players spawn")
+
 def create_random_actions():
     """ Return random actions """
     ret = [[],[]]
@@ -28,12 +39,12 @@ def create_random_actions():
             ret[plridx].append(r.randint(1,4))
     return ret
     
-def run_async_torille(tick_counter, quit_flag, match_frames, turn_frames):
+def run_async_torille(tick_counter, quit_flag, match_frames, turn_frames, engagement_distance):
     # Runs Toribash and increments the Value tick_counter on every frame
     controller = ToribashControl()
     controller.settings.set("matchframes", match_frames)
     controller.settings.set("turnframes", turn_frames)
-    controller.settings.set("engagement_distance", 1500)
+    controller.settings.set("engagement_distance", engagement_distance)
     controller.init()
     while quit_flag.value == 0:
         s,terminal = controller.get_state()
@@ -46,7 +57,10 @@ def run_async_torille(tick_counter, quit_flag, match_frames, turn_frames):
     controller.close()
     
 def test_async(num_instances, warm_up_seconds, benchmark_seconds, match_frames,
-               turn_frames):
+               turn_frames, engagement_distance):
+    """ 
+    Run benchmark with given parameters, returns [turns per second, frames per second]
+    """
     last_ticks = [0 for i in range(num_instances)]
     tick_ctrs = [Value("i") for i in range(num_instances)]
     quit_flags = [Value("i") for i in range(num_instances)]
@@ -55,7 +69,8 @@ def test_async(num_instances, warm_up_seconds, benchmark_seconds, match_frames,
         process = Process(target=run_async_torille, args=(tick_ctrs[i],
                                                           quit_flags[i],
                                                           match_frames,
-                                                          turn_frames))
+                                                          turn_frames, 
+                                                          engagement_distance))
         process.start()
         runners.append(process)
     
@@ -75,27 +90,28 @@ def test_async(num_instances, warm_up_seconds, benchmark_seconds, match_frames,
         ticks_progressed += ticks[i]-start_ticks[i]
     pps = ticks_progressed / benchmark_seconds
     
-    print("PPS: %.2f" % pps)
-    print("FPS: %.2f\n" % (pps*turn_frames))
-    
     for i in range(num_instances):
         quit_flags[i].value = 1
         runners[i].join()
+
+    return [pps, (pps*turn_frames)]
     
 if __name__ == '__main__':
-    if len(sys.argv) != 4:
-        print("Usage: python3 async_benchmark.py num_instances match_frames turn_frames")
-    else:
-        num_instances = int(sys.argv[1])
-        match_frames = int(sys.argv[2])
-        turn_frames = int(sys.argv[3])
-        print(("Instances: %d\nWarmup: %d s\nBenchmark: %d s" +
-              "\nMatchframes: %d\nTurnframes: %d") %
-              (num_instances, WARM_UP_SECONDS, BENCHMARK_SECONDS, 
-               match_frames, turn_frames))
-        test_async(num_instances=num_instances,
-                   warm_up_seconds=WARM_UP_SECONDS,
-                   benchmark_seconds=BENCHMARK_SECONDS,
-                   match_frames=match_frames,
-                   turn_frames=turn_frames)
+    args = parser.parse_args()
+    
+    assert args.turn_frames > 1
+    assert args.turn_frames < args.match_frames
+    assert args.engagement_distance > 0 
 
+    [pps, fps] = test_async(num_instances=args.num_instances,
+                            warm_up_seconds=args.warmup_time,
+                            benchmark_seconds=args.benchmark_time,
+                            match_frames=args.match_frames,
+                            turn_frames=args.turn_frames,
+                            engagement_distance=args.engagement_distance)
+    print(("Instances: %d\nWarmup: %d s\nBenchmark: %d s" +
+              "\nMatchframes: %d\nTurnframes: %d\nEngagement distance: %d") %
+              (args.num_instances, args.warmup_time, args.benchmark_time, 
+               args.match_frames, args.turn_frames, args.engagement_distance))
+    print("PPS: %.2f" % pps)
+    print("FPS: %.2f\n" % fps)