+++ torille/torille.py
@@ -33,6 +33,7 @@ import pprint
 from filelock import FileLock
 import warnings
 from copy import deepcopy
+from stat import S_IREAD, S_IRGRP, S_IROTH
 
 def create_random_actions():
     """ Return random actions for ToribashControl """
@@ -41,6 +42,26 @@ def create_random_actions():
         for jointidx in range(ToribashConstants.NUM_CONTROLLABLES):
             ret[plridx].append(r.randint(1,4))
     return ret
+
+def set_file_readonly(filepath):
+    """
+    Attempt to set given file read-only, and
+    return True on success.
+
+    Parameters:
+        filepath: Path to file to be set read-only
+    Returns:
+        success: True if file was set read only, otherwise False
+    """
+    if os.path.isfile(filepath):
+        # Set to read only (for user, group and all)
+        try:
+            os.chmod(self.toribash_stderr_file, S_IREAD | S_IRGRP | S_IROTH)
+        except Exception as e:
+            return False
+        return True
+    else:
+        return False
 
 def check_darwin_sanity():
     """ 
@@ -131,47 +152,103 @@ class ToribashConstants:
     # Number of setting variables
     NUM_SETTINGS = 19
 
-    # Bodypart x,y,z + Joint states + hand grips + injuries
-    STATE_LENGTH = (NUM_LIMBS*3*2) + NUM_JOINTS*2 + 4 + 2
+    # Bodypart x,y,z + Bodypart x,y,z velocities + 
+    # groin rotation + Joint states + hand grips + injuries
+    STATE_LENGTH = (NUM_LIMBS*3*2)*2 + 16*2 + NUM_JOINTS*2 + 4 + 2
 
     # Path to Toribash supplied with the wheel package
     # This should be {this file}/toribash/toribash.exe
     my_dir = os.path.dirname(os.path.realpath(__file__))
     TORIBASH_EXE = os.path.join(my_dir, "toribash", "toribash.exe")
 
+    # Path to Toribash's stderr.txt file, which
+    # will be filled with bunch of errors unless handled separately
+    TORIBASH_STDERR_FILE = os.path.join(my_dir, "toribash", "stderr.txt")
+
 class ToribashState:
     """ 
     Class for storing and processing the state representations
     from Toribash 
     """
-    def __init__(self, state):
+    def __init__(self, state, winner=None):
         # Limb locations
         # For both players, for all limbs, x,y,z coordinates
         self.limb_positions = np.zeros((2,ToribashConstants.NUM_LIMBS,3))
+        # Limb velocities
+        # For both players, for all limbs, x,y,z velocities
+        self.limb_velocities = np.zeros((2,ToribashConstants.NUM_LIMBS,3))
+        # Groin rotations of both players
+        # Rotation is defined as 4x4 rotation matrix
+        self.groin_rotations = np.zeros((2, 4, 4))
         # Joint states (including hands)
         # For both players
         self.joint_states = np.zeros((2,ToribashConstants.NUM_CONTROLLABLES))
         # Amount of injury of players 
         # For both players
         self.injuries = np.zeros((2,))
+        # Winner of the game (only defined at end of the games)
+        # 0 = tie, 1 = player 1 won, 2 = player 2 won
+        self.winner = winner
         
         self.process_list(state)
         
     def process_list(self, state_list):
-        """ Updates state representations according to given list of 
-        variables from Toribash """
+        """ 
+        Updates state representations according to given list of 
+        variables from Toribash
+        """
         # Indexes from  state_structure.md
         # Limbs
         self.limb_positions[0] = np.array(state_list[:63]).reshape(
                                         (ToribashConstants.NUM_LIMBS,3))
-        self.limb_positions[1] = np.array(state_list[86:149]).reshape(
+        self.limb_velocities[0] = np.array(state_list[63:126]).reshape(
                                         (ToribashConstants.NUM_LIMBS,3))
+        self.groin_rotations[0] = np.array(state_list[126:142]).reshape(4,4)
+        
+        self.limb_positions[1] = np.array(state_list[165:228]).reshape(
+                                        (ToribashConstants.NUM_LIMBS,3))
+        self.limb_velocities[1] = np.array(state_list[228:291]).reshape(
+                                        (ToribashConstants.NUM_LIMBS,3))
+        self.groin_rotations[1] = np.array(state_list[291:307]).reshape(4,4)
+        
         # Joint states (inc. hand grips)
-        self.joint_states[0] = np.array(state_list[63:85], dtype=np.int)
-        self.joint_states[1] = np.array(state_list[149:171], dtype=np.int)
+        self.joint_states[0] = np.array(state_list[142:164], dtype=np.int)
+        self.joint_states[1] = np.array(state_list[307:329], dtype=np.int)
         # Injuries
-        self.injuries[0] = state_list[85]
-        self.injuries[1] = state_list[171]
+        self.injuries[0] = state_list[164]
+        self.injuries[1] = state_list[329]
+
+    def get_normalized_locations(self):
+        """
+        Normalizes and returns limb locations which are centered
+        around respective player's groin, and applies groin's 
+        rotation to the locations.
+
+        Applies following operations in order:
+            - limb_locations - location of player's groin
+            - Apply rotation player's groin to centered coordinates
+
+        E.g. at the start of game both players will have same 
+             coordinates from their point of view.
+
+        Returns:
+            normalized_limb_positions: A (2, 2, NUM_LIMBS, 3) array
+                                       of normalized locations, from the
+                                       point-of-view of both players.
+        """
+
+        # Body-part 4 is "groin"
+        # Center around the local-player's groin
+        player1_obs = self.limb_positions - self.limb_positions[0,4]
+        player2_obs = self.limb_positions - self.limb_positions[1,4]
+        
+        # Apply rotation of the groin, otherwise
+        # player2 will have "mirrored" coordinates
+        rotations = self.groin_rotations[:, :3, :3]
+        player1_obs = np.dot(player1_obs.reshape((-1, 3)), rotations[0]).reshape((2, ToribashConstants.NUM_LIMBS, 3))
+        player2_obs = np.dot(player2_obs.reshape((-1, 3)), rotations[1]).reshape((2, ToribashConstants.NUM_LIMBS, 3))
+
+        return np.array((player1_obs, player2_obs))
 
 class ToribashSettings:
     """ Class for storing and processing settings for Toribash """
@@ -261,7 +338,7 @@ class ToribashControl:
     def __init__(self, 
                  settings=None, 
                  draw_game=False,
-                 executable=ToribashConstants.TORIBASH_EXE, 
+                 executable=ToribashConstants.TORIBASH_EXE,
                  port=ToribashConstants.PORT):
         """ 
         Parameters:
@@ -281,6 +358,12 @@ class ToribashControl:
         if not os.path.isfile(self.executable_path):
             raise ValueError("Toribash executable path is not a file: %s" % 
                              self.executable_path)
+        # Create path to stderr.txt file which is created by the 
+        # Toribash executable (next to it)
+        self.toribash_stderr_file = os.path.join(
+            os.path.dirname(executable), "stderr.txt"
+        )
+
         self.process = None
         self.connection = None
         self.port = port
@@ -318,6 +401,9 @@ class ToribashControl:
             if sys.platform == "linux":
                 # Sanity check launching on Linux
                 check_linux_sanity()
+                # Attempt to make stderr.txt read-only
+                # TODO this will cause headache when trying to remove torille
+                _ = set_file_readonly(self.toribash_stderr_file)
                 # Add wine command explicitly for running on Linux
                 self.process = subprocess.Popen(("nohup", "wine", self.executable_path), 
                                              stdout=subprocess.DEVNULL, 
@@ -325,6 +411,8 @@ class ToribashControl:
             elif sys.platform == "darwin":
                 # Sanity check launching on OSX
                 check_darwin_sanity()
+                # Attempt to make stderr.txt read-only
+                _ = set_file_readonly(self.toribash_stderr_file)
                 # Add wine command for running on osx
                 self.process = subprocess.Popen(("wine %s" % self.executable_path), 
                                              stdout=subprocess.DEVNULL, 
@@ -332,6 +420,8 @@ class ToribashControl:
                                              shell=True)
             else:
                 # Launch on Windows (just call the .exe)
+                # Don't try to set stderr.txt to read-only: This will
+                # cause Toribash to crash on Windows
                 self.process = subprocess.Popen((self.executable_path,), 
                                                 stdout=subprocess.DEVNULL, 
                                                 stderr=subprocess.DEVNULL)
@@ -353,10 +443,10 @@ class ToribashControl:
             conn.settimeout(ToribashConstants.TIMEOUT)
             self.connection = conn
         # Send handshake 
-        self._send_comma_list([int(self.draw_game)])
+        self._send_comma_list(self.connection, [int(self.draw_game)])
         # Send initial settings
         self.settings.validate_settings()
-        self._send_comma_list(self.settings.settings)
+        self._send_comma_list(self.connection, self.settings.settings)
 
     def close(self):
         """ Close the running Toribash instance and clean up """
@@ -385,9 +475,14 @@ class ToribashControl:
         """
         s = self._recv_line(self.connection).decode()
         terminal = s.startswith("end")
+        winner = None
         if terminal:
-            # Remove first three characters + comma to parse the state
-            s = s[4:]
+            # After 'end' comes ':#' where # specifies the winner
+            # (one-digit integer) 
+            # Read the winner
+            winner = int(s[4])
+            # Remove first three characters + double-dots + integer + comma
+            s = s[6:]
             # Allow calling reset next
             self.requires_reset = True
         s = list(map(float, s.split(",")))
@@ -396,17 +491,18 @@ class ToribashControl:
             raise ValueError(("Got state of invalid size. Expected %d, got %d"+
                              "\nState: %s") %
                              (ToribashConstants.STATE_LENGTH, len(s), s))
-        return s, terminal
+        return s, terminal, winner
         
-    def _send_comma_list(self, data):
+    def _send_comma_list(self, s, data):
         """ 
         Send given list to Toribash as comma-separated list
         Parameters:
+            s: The socket where to send the data
             data: List of values to be sent
         """
         # We need to add end of line for the luasocket "*l"
         data = ",".join(map(str, data)) + "\n"
-        self.connection.sendall(data.encode())
+        s.sendall(data.encode())
         
     def get_state(self):
         """ 
@@ -417,8 +513,8 @@ class ToribashControl:
         """
         self._check_if_initialized()
         
-        s, terminal = self._recv_state()
-        s = ToribashState(s)
+        s, terminal, winner = self._recv_state()
+        s = ToribashState(s, winner)
         return s, terminal
     
     def reset(self):
@@ -437,7 +533,7 @@ class ToribashControl:
         # Validate settings
         self.settings.validate_settings()
 
-        self._send_comma_list(self.settings.settings)
+        self._send_comma_list(self.connection, self.settings.settings)
         s,terminal = self.get_state()
         self.requires_reset = False
         return s
@@ -472,7 +568,8 @@ class ToribashControl:
             # Check both players at the same time
             if (actions[0][i] > 4 or actions[0][i] < 1 or actions[1][i] > 4 or
                     actions[1][i] < 1):
-                raise ValueError("Joint states should be in {1,2,3,4}")
+                raise ValueError("Joint states should be in {1,2,3,4}. "+
+                    "Note: Gym environments take in {0,1,2,3}")
     
     def make_actions(self, actions):
         """ 
@@ -504,7 +601,7 @@ class ToribashControl:
         # Concat lists into one 
         actions = actions[0]+actions[1]
 
-        self._send_comma_list(actions)
+        self._send_comma_list(self.connection, actions)
     
     def get_state_dim(self):
         """ Return size of state space per character """